Add files via upload
The code is of black scholes model specifically written for Bitcoin. What is the black scholes model? The Black-Scholes model is a differential equation used to estimate the fair value of an options contract. It helps traders determine whether an option is "cheap" or "expensive" based on market conditions. It revolutionized modern finance and won its creators the Nobel Prize in Economics.  FeaturesPrice Calculation: Calculates the theoretical price for Call and Put options.The Greeks: Computes risk sensitivities:$\Delta$ Delta: Sensitivity to asset price.$\Gamma$ Gamma: Rate of change of Delta.$\Theta$ Theta: Sensitivity to time decay.$\nu$ Vega: Sensitivity to volatility.$\rho$ Rho: Sensitivity to interest rates. You can copy the code and run it and change the inputs according to options. Good Luck Trading!